# Conflicts:
#	nts.uk/web/nts.uk.com.web/src/main/webapp/lib/nittsu/time.js.map
#	nts.uk/web/nts.uk.com.web/src/main/webapp/lib/nittsu/ui/ko-ext/datepicker-ko-ext.js.map
#	nts.uk/web/nts.uk.com.web/src/main/webapp/lib/nittsu/util.js
#	nts.uk/web/nts.uk.com.web/src/main/webapp/lib/nittsu/util.js.map

+++ nts.uk/web/nts.uk.pr.web/src/main/webapp/view/qmm/012/b/viewmodel.ts
@@ -29,6 +29,7 @@ module qmm012.b.viewmodel {
         texteditor_B_INP_004: any;
         enable_B_INP_002: KnockoutObservable<boolean> = ko.observable(false);
         screenModel: qmm012.b.ScreenModel;
+        B_BTN_004_enable: KnockoutObservable<boolean> = ko.observable(true);
         constructor(screenModel: qmm012.b.ScreenModel) {
             let self = this;
             self.screenModel = screenModel;
@@ -46,31 +47,39 @@ module qmm012.b.viewmodel {
                 let categoryAtr;
                 switch (newValue) {
                     case 1:
+                        //select  all  
                         categoryAtr = -1
                         break;
                     case 2:
+                        // 支給
                         categoryAtr = 0
                         break;
                     case 3:
+                        // 控除
                         categoryAtr = 1
                         break;
                     case 4:
+                        // 勤怠
                         categoryAtr = 2
                         break;
                     case 5:
+                        //記事
                         categoryAtr = 3
                         break;
                     case 6:
+                        //その他
                         categoryAtr = 9
                         break;
                 }
                 self.categoryAtr = categoryAtr;
+                //then load gridlist
                 self.LoadGridList();
             });
             self.checked_B_002.subscribe(function() {
                 self.LoadGridList();
             })
             // set gridlist data
+            //gridlist column
             self.GridColumns_B_001 = ko.observableArray([
                 { headerText: '項目区分', prop: 'categoryAtrName', width: 80 },
                 { headerText: 'コード', prop: 'itemCode', width: 50 },
@@ -80,14 +89,18 @@ module qmm012.b.viewmodel {
 
             function makeIcon(val) {
                 if (val == 1)
+                    //it  mean この項目名を廃止する , bind X icon
                     return "<div class = 'NoIcon' > </div>";
                 return "";
             }
             self.GridlistCurrentCode_B_001.subscribe(function(newValue) {
+                //get ItemModel by itemCode has selected
                 var item = _.find(self.GridlistItems_B_001(), function(ItemModel: service.model.ItemMaster) {
                     return ItemModel.itemCode == newValue;
                 });
+                //set it to GridlistCurrentItem_B_001 for trigger subscribe
                 self.GridlistCurrentItem_B_001(item);
+                //set text for B_INP_002
                 self.B_INP_002_text(newValue);
             });
 
@@ -95,14 +108,14 @@ module qmm012.b.viewmodel {
                 self.GridCurrentItemName_B_001(itemModel ? itemModel.itemName : '');
                 self.GridCurrentUniteCode_B_001(itemModel ? itemModel.uniteCode : '');
                 self.GridCurrentCategoryAtr_B_001(itemModel ? itemModel.categoryAtr : 0);
-                //Because there are many items in the same group  After set value , need call ChangeGroup function for Set Value to layout
+                // If set same group, GridCurrentCategoryAtr_B_001 will not trigger subscribe => can't call ChangeGroup function for load that item data so call ChangeGroup function here
                 ChangeGroup(self.GridCurrentCategoryAtr_B_001());
                 self.GridCurrentCodeAndName_B_001(itemModel ? itemModel.itemCode + ' ' + itemModel.itemName : '');
                 self.GridCurrentDisplaySet_B_001(itemModel ? itemModel.displaySet == 1 ? true : false : false);
                 self.GridCurrentItemAbName_B_001(itemModel ? itemModel.itemAbName : '');
                 self.GridCurrentCategoryAtrName_B_001(itemModel ? itemModel.categoryAtrName : '');
-                //when CurrentCode != undefined , need disable INP_002
-                if (self.GridlistCurrentCode_B_001() != undefined) {
+                //when itemModel != undefined , need disable INP_002
+                if (itemModel != undefined) {
                     self.enable_B_INP_002(false);
                 }
             });
@@ -119,26 +132,39 @@ module qmm012.b.viewmodel {
                 $('#screenF').hide();
                 switch (newValue) {
                     case 0:
+                        //支給
                         $('#screenC').show();
                         self.screenModel.screenModelC.CurrentItemMaster(self.GridlistCurrentItem_B_001());
                         break;
                     case 1:
+                        //控除
                         $('#screenD').show();
                         self.screenModel.screenModelD.CurrentItemMaster(self.GridlistCurrentItem_B_001());
                         break;
                     case 2:
+                        //勤怠
                         $('#screenE').show();
                         self.screenModel.screenModelE.CurrentItemMaster(self.GridlistCurrentItem_B_001());
                         break;
                     case 3:
+                        //記事
                         $('#screenF').show();
                         self.screenModel.screenModelF.CurrentItemMaster(self.GridlistCurrentItem_B_001());
                         break;
                 }
+                //why don't have その他 ? because it show nothing
             }
-
+            //first load , need call LoadGridList
             self.LoadGridList();
-
+            self.enable_B_INP_002.subscribe(function(newValue) {
+                if (newValue) {
+                    //it mean new item mode
+                    self.setNewItemMode();
+                } else {
+                    //it mean update item mode
+                    self.setUpdateItemMode();
+                }
+            })
             //set text editer data
             //INP_002
             self.texteditor_B_INP_002 = {
@@ -169,10 +195,34 @@ module qmm012.b.viewmodel {
                 }))
             };
         }
-
-        GetCurrentItemMaster() {
+        setNewItemMode() {
             let self = this;
-            let CurrentGroup = self.GridCurrentCategoryAtr_B_001();
+            //set selected code is '' for trigger subscribe
+            self.GridlistCurrentCode_B_001('');
+            //disable delete button
+            self.B_BTN_004_enable(false);
+            //disable 有効期間設定 button
+            self.screenModel.screenModelC.C_BTN_001_enable(false);
+            self.screenModel.screenModelC.C_BTN_002_enable(false);
+            self.screenModel.screenModelD.D_BTN_001_enable(false);
+            self.screenModel.screenModelD.D_BTN_002_enable(false);
+        }
+        setUpdateItemMode() {
+            let self = this;
+            //if from new mode change to update mode , need clear all ntsError 
+            $('#B_INP_002').ntsError('clear');
+            //enable delete button
+            self.B_BTN_004_enable(true);
+            //enable 有効期間設定 button
+            self.screenModel.screenModelC.C_BTN_001_enable(true);
+            self.screenModel.screenModelC.C_BTN_002_enable(true);
+            self.screenModel.screenModelD.D_BTN_001_enable(true);
+            self.screenModel.screenModelD.D_BTN_002_enable(true);
+        }
+        GetCurrentItemMaster() {
+            //get item master The user entered on the form 
+            let self = this;
+            //this is item master Constructor
             let itemMaster = new service.model.ItemMaster(
                 self.B_INP_002_text(),
                 self.GridCurrentItemName_B_001(),
@@ -187,6 +237,7 @@ module qmm012.b.viewmodel {
                 self.getCurrentItemDisplayAtr(),
                 1
             )
+            //set sub item constructor
             itemMaster.itemSalary = self.screenModel.screenModelC.GetCurrentItemSalary();
             itemMaster.itemDeduct = self.screenModel.screenModelD.GetCurrentItemDeduct();
             itemMaster.itemAttend = self.screenModel.screenModelE.getCurrentItemAttend();
@@ -196,44 +247,59 @@ module qmm012.b.viewmodel {
         LoadGridList(ItemCode?) {
             let self = this;
             let categoryAtr = self.categoryAtr;
+            //load dispSet 
+            //if 0  mean
+            // no view この項目名を廃止する 
+            //else view all
             let dispSet = self.checked_B_002() ? -1 : 0;
+            //call service load findAllItemMaster
             service.findAllItemMaster(categoryAtr, dispSet).done(function(MasterItems: Array<service.model.ItemMaster>) {
                 self.GridlistItems_B_001(MasterItems);
                 //set selected first item in list
                 if (self.GridlistItems_B_001().length > 0)
                     // if not itemcode parameter
                     if (!ItemCode)
+                        //set GridlistCurrentCode_B_001 selected first item in gridlist
                         self.GridlistCurrentCode_B_001(self.GridlistItems_B_001()[0].itemCode);
                     else
+                        //set  selected == param itemcode
                         self.GridlistCurrentCode_B_001(ItemCode);
             }).fail(function(res) {
                 alert(res);
             });
         }
 
-        DeleteDialog() {
+        deleteItem() {
             let self = this;
             let ItemMaster = self.GetCurrentItemMaster();
             let index = self.GridlistItems_B_001.indexOf(self.GridlistCurrentItem_B_001());
-            service.deleteItemMaster(ItemMaster).done(function(any) {
-                //reload grid and set select code after delete item success
-                if (index) {
-                    let selectItemCode;
-                    if (self.GridlistItems_B_001().length - 1 > 1) {
-                        if (index < self.GridlistItems_B_001().length - 1)
-                            selectItemCode = self.GridlistItems_B_001()[index - 1].itemCode;
-                        else
-                            selectItemCode = self.GridlistItems_B_001()[index - 2].itemCode;
-                    } else
-                        selectItemCode = '';
-                    self.LoadGridList(selectItemCode);
-                }
-            }).fail(function(res) {
-
-                alert(res);
-            });
+            //if has item selected
+            if (index >= 0) {
+                //show dialog
+                nts.uk.ui.dialog.confirm("データを削除します。\r\nよろしいですか？").ifYes(function() {
+                    //if yes call service delete item
+                    service.deleteItemMaster(ItemMaster).done(function(any) {
+                        //reload grid and set select code after delete item success
+                        let selectItemCode;
+                        //if after delete gridlist length >0
+                        if (self.GridlistItems_B_001().length - 1 > 1) {
+                            if (index < self.GridlistItems_B_001().length - 1)
+                                //if not last selected item , set selected same position
+                                selectItemCode = self.GridlistItems_B_001()[index - 1].itemCode;
+                            else
+                                //else selected item Before it
+                                selectItemCode = self.GridlistItems_B_001()[index - 2].itemCode;
+                        } else
+                            //length < 0 no select any thing
+                            selectItemCode = '';
+                        //reload gruid list
+                        self.LoadGridList(selectItemCode);
+                    }).fail(function(res) {
+                        alert(res);
+                    });
+                });
+            }
         }
-
 
         getCurrentZeroDisplaySet() {
             let Result;
@@ -241,18 +307,27 @@ module qmm012.b.viewmodel {
             let CurrentGroup = self.GridCurrentCategoryAtr_B_001();
             switch (CurrentGroup) {
                 case 0:
+                    //支給
                     Result = self.screenModel.screenModelC.CurrentZeroDisplaySet();
                     break;
                 case 1:
+                    //控除
                     Result = self.screenModel.screenModelD.CurrentZeroDisplaySet();
                     break;
                 case 2:
+                    //勤怠
                     Result = self.screenModel.screenModelE.CurrentZeroDisplaySet();
                     break;
+                //記事
+                case 3:
+                    Result = self.screenModel.screenModelF.CurrentZeroDisplaySet();
+                    break;
+
             }
             return Result;
         }
         getCurrentItemDisplayAtr() {
+            //like getCurrentZeroDisplaySet
             let Result;
             let self = this;
             let CurrentGroup = self.GridCurrentCategoryAtr_B_001();
@@ -266,18 +341,22 @@ module qmm012.b.viewmodel {
                 case 2:
                     Result = self.screenModel.screenModelE.CurrentItemDisplayAtr();
                     break;
+                case 3:
+                    Result = self.screenModel.screenModelF.CurrentItemDisplayAtr();
+                    break;
             }
             return Result;
         }
         openADialog() {
+            //open select type dialog for new item
             let self = this;
             nts.uk.ui.windows.sub.modal('../a/index.xhtml', { height: 480, width: 630, dialogClass: "no-close" }).onClosed(function(): any {
                 if (nts.uk.ui.windows.getShared('groupCode') != undefined) {
+                    //get group from session
                     let groupCode = Number(nts.uk.ui.windows.getShared('groupCode'));
                     //set layout for new.
-                    self.GridlistCurrentCode_B_001('');
-                    self.GridCurrentCategoryAtr_B_001(groupCode);
                     self.enable_B_INP_002(true);
+                    self.GridCurrentCategoryAtr_B_001(groupCode);
                 }
             });
         }
@@ -288,12 +367,15 @@ module qmm012.b.viewmodel {
             if (self.enable_B_INP_002()) {
                 self.addNewItemMaster(ItemMaster);
             } else {
+                //else update mode
                 self.updateItemMaster(ItemMaster);
             }
         }
         addNewItemMaster(ItemMaster: service.model.ItemMaster) {
             let self = this;
+            //call add service
             service.addItemMaster(ItemMaster).done(function(any) {
+                //after add , reload grid list
                 self.LoadGridList(ItemMaster.itemCode);
             }).fail(function(res) {
                 alert(res);
@@ -301,14 +383,14 @@ module qmm012.b.viewmodel {
         }
         updateItemMaster(ItemMaster: service.model.ItemMaster) {
             let self = this;
+            //call update service
             service.updateItemMaster(ItemMaster).done(function(any) {
+                //after add , reload grid list
                 self.LoadGridList(ItemMaster.itemCode);
             }).fail(function(res) {
                 alert(res);
             });
-
         }
-
         openJDialog() {
             let self = this;
             nts.uk.ui.windows.sub.modal('../j/index.xhtml', { height: 700, width: 970, dialogClass: "no-close" }).onClosed(function(): any {
